pyproject.toml
[tool.poetry]
name = "diploma-2024"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10,<3.13"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
pyautogen = "^0.2.19"
aiohttp = "^3.9.3"
pydantic-settings = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
